Resize 32bit

+++ include/SmithWaterman.hpp
@@ -20,17 +20,17 @@ class SmithWaterman {
   SmithWaterman(const std::string& query_seq_path,
                 const std::string& target_seq_path);
 
-  std::vector<size_t> solve();
+  std::vector<uint32_t> solve();
 
   int validate(const std::string& ref_path);
 
   void report() const;
 
-  static constexpr int64_t match_score = 5;      // Do not modify.
-  static constexpr int64_t mismatch_score = -3;  // Do not modify.
-  static constexpr int64_t gap_score = -4;       // Do not modify.
+  static constexpr int32_t match_score = 5;      // Do not modify.
+  static constexpr int32_t mismatch_score = -3;  // Do not modify.
+  static constexpr int32_t gap_score = -4;       // Do not modify.
   // The highest scores
-  std::vector<size_t> max_scores;
+  std::vector<uint32_t> max_scores;
 
  private:
 
@@ -39,23 +39,23 @@ class SmithWaterman {
   std::vector<FastaSequence> target_seqs;
 
   // The reference scores.
-  std::vector<size_t> refs;
+  std::vector<uint32_t> refs;
 
-  std::vector<std::pair<size_t,size_t> > problems;
+  std::vector<std::pair<uint32_t,uint32_t> > problems;
 
   // Number of query/target sequences
-  size_t query_seqs_size;
-  size_t target_seqs_size;
+  uint32_t query_seqs_size;
+  uint32_t target_seqs_size;
 
   // Length of each sequence
-  //std::vector<std::vector<int64_t>> query_seqs_lens;
-  //std::vector<std::vector<int64_t>> target_seqs_lens;
+  //std::vector<std::vector<int32_t>> query_seqs_lens;
+  //std::vector<std::vector<int32_t>> target_seqs_lens;
 
   // The indices of the highest score for each query-target pair.
-  //std::vector<size_t> max_positions;
+  //std::vector<uint32_t> max_positions;
 
   void read_seq(const std::string& seq_path, std::vector<FastaSequence>& seqs);
 
-  void read_ref(const std::string& ref_path, std::vector<size_t>& refs);
+  void read_ref(const std::string& ref_path, std::vector<uint32_t>& refs);
 };
-void pair_align(FastaSequence* query_seq, FastaSequence* target_seq, size_t* sw);
+void pair_align(FastaSequence* query_seq, FastaSequence* target_seq, uint32_t* sw);
